move shared_utils to src/utils
