+++ backend/src/utils/shared_utils/index.js
